// apps/web/src/domains/canvas-management/frontend/components/react-flow-wrapper/components/multi-select/selection-bounding-box/core/types.ts
export interface BoundingBoxBounds {
  left: number;
  top: number;
  width: number;
  height: number;
}